api/requirements.txt
fastapi==0.115.8
uvicorn==0.33.0
SQLAlchemy==2.0.37
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose==3.3.0
httpx==0.28.1
email-validator==2.2.0
